# optimize DELTA_BYTE_ARRAY decoding